# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dtos.CreateMembershipDTO:
    properties:
      company_id:
        description: Nullable para SuperAdmin
        type: integer
      invited_by:
        type: integer
      is_default:
        type: boolean
      role:
        enum:
        - superadmin
        - admin
        - recruiter
        - user
        type: string
      status:
        enum:
        - active
        - inactive
        - pending
        type: string
      user_id:
        minimum: 1
        type: integer
    required:
    - role
    - user_id
    type: object
...
  dtos.CreateSystemValueDTO:
    properties:
      category:
        type: string
      company_id:
        type: integer
      description:
        type: string
      display_order:
        type: integer
      label:
        type: string
      value:
        type: string
    required:
    - category
    - label
    - value
    type: object
...
host: localhost:8001
info:
  contact:
    email: support@dvra.com
    name: API Support
  description: API para sistema de reclutamiento y gestión de candidatos
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: DVRA API
  version: 1.2.0
paths:
...
  /admin/memberships:
    post:
      consumes:
      - application/json
      description: Asigna un usuario a una empresa (solo SuperAdmin)
      parameters:
      - description: Datos de la membresía
        in: body
        name: membership
        required: true
        schema:
          $ref: '#/definitions/dtos.CreateMembershipDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Crear membresía
      tags:
      - Memberships
...
  /admin/system-values:
    get:
      consumes:
      - application/json
      description: Obtiene todos los valores de sistema (solo SuperAdmin)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Listar todos los valores de sistema
      tags:
      - System Values
    post:
      consumes:
      - application/json
      description: Crea un nuevo valor de sistema (solo SuperAdmin)
      parameters:
      - description: Datos del valor
        in: body
        name: value
        required: true
        schema:
          $ref: '#/definitions/dtos.CreateSystemValueDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Crear valor de sistema
      tags:
      - System Values
...
  /memberships:
    get:
      consumes:
      - application/json
      description: Obtiene membresías (todas si es SuperAdmin, de la empresa si es
        usuario normal)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Listar membresías
      tags:
      - Memberships
...
  /system-values/{category}:
    get:
      consumes:
      - application/json
      description: Retorna todos los valores de sistema de una categoría específica
      parameters:
      - description: 'Categoría (ej: employment_type, contract_type)'
        in: path
        name: category
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Obtener valores por categoría
      tags:
      - System Values
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"